+++ lib/src/service/tweets/tweets_service.dart
@@ -29,6 +29,7 @@ import 'tweet_count_data.dart';
 import 'tweet_count_granularity.dart';
 import 'tweet_count_meta.dart';
 import 'tweet_data.dart';
+import 'tweet_exclude_type.dart';
 import 'tweet_expansion.dart';
 import 'tweet_field.dart';
 import 'tweet_geo_param.dart';
@@ -1855,6 +1856,12 @@ abstract class TweetsService {
   ///                   [untilTweetId]  will be forced to the most recent ID
   ///                   available.
   ///
+  /// - [excludes]: The list of the types of Tweets to exclude from the
+  ///               response. When exclude=retweets is used, the maximum
+  ///               historical Tweets returned is still 3200. When the
+  ///               exclude=replies parameter is used for any value, only the
+  ///               most recent 800 Tweets are available.
+  ///
   /// - [expansions]: Expansions enable you to request additional data objects
   ///                 that relate to the originally returned Tweets. Submit a
   ///                 list of desired expansions in a comma-separated list
@@ -1949,6 +1956,7 @@ abstract class TweetsService {
     DateTime? endTime,
     String? sinceTweetId,
     String? untilTweetId,
+    List<TweetExcludeType>? excludes,
     List<TweetExpansion>? expansions,
     List<TweetField>? tweetFields,
     List<UserField>? userFields,
@@ -2969,6 +2977,7 @@ class _TweetsService extends BaseService implements TweetsService {
     DateTime? endTime,
     String? sinceTweetId,
     String? untilTweetId,
+    List<TweetExcludeType>? excludes,
     List<TweetExpansion>? expansions,
     List<TweetField>? tweetFields,
     List<UserField>? userFields,
@@ -2987,6 +2996,7 @@ class _TweetsService extends BaseService implements TweetsService {
             'end_time': endTime,
             'since_id': sinceTweetId,
             'until_id': untilTweetId,
+            'exclude': excludes,
             'expansions': expansions,
             'tweet.fields': tweetFields,
             'user.fields': userFields,