+++ CHANGELOG.md
@@ -48,6 +48,7 @@
   - GET /2/tweets/counts/all
   - GET /2/tweets/counts/recent
 - Added `sinceTweetId` and `untilTweetId` parameters to `TweetsService.lookupTweets`. ([#309](https://github.com/twitter-dart/twitter-api-v2/issues/309))
+- Added `excludes` parameter to `TweetsService.lookupTweets`. ([#308](https://github.com/twitter-dart/twitter-api-v2/issues/308))
 
 ## v2.7.0